Port `remapKotlinModule`
