--- src/main/kotlin/com/xpdustry/ksr/KotlinShadowRelocatorPlugin.kt
@@ -1,43 +1,0 @@
-/*
- * This file is part of KSR, a gradle plugin for handling Kotlin metadata relocation.
- *
- * MIT License
- *
- * Copyright (c) 2025 Xpdustry
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package com.xpdustry.ksr
-
-import com.github.jengelman.gradle.plugins.shadow.ShadowJavaPlugin
-import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-
-internal class KotlinShadowRelocatorPlugin : Plugin<Project> {
-  override fun apply(target: Project) {
-    target.plugins.withType(ShadowJavaPlugin::class.java).configureEach {
-      target.tasks.withType(ShadowJar::class.java).configureEach { task ->
-        task.doLast("relocateMetadata") {
-          relocateMetadata(task)
-        }
-      }
-    }
-  }
-}
